railmadadBackend/requirements.txt
Django>=5.0
Pillow>=9.0
torch>=2.0
transformers>=4.44.0
moviepy>=0.4.51
google-generativeai>=0.7.0
django-environ>=0.11.2
djangorestframework>=3.15.0
django-cors-headers>=4.0.0